+++ zeldarose/train_transformer.py
@@ -279,7 +279,7 @@ class SavePretrainedModelCallback(pl.callbacks.Callback):
     help=(
         "A raw corpus for validation."
         " Either as a path or as a `handle:config:split` identifier for 🤗 hub."
-        " (handle can be a url)"
+        " (handle can be a url), e.g. `lgrobol/openminuscule:text:train`"
     ),
 )
 @click.option("--verbose", is_flag=True, help="More detailed logs")